kc-4 refactor
