+++ src/action/setup/webcam/index.js
@@ -1,12 +1,12 @@
 import { store } from '../../../model/store';
 import { WebCamConstants } from '../../../model/constant'
-import { webcamStreamSetup } from '../../webcam-stream-setup'
-import { getCameras } from '../../get-cameras'
+import { setWebcamStream } from '../../webcam/set-webcam-stream'
+import { getCameras } from '../../webcam/get-cameras'
 
 export const setUpWebCamera = () => {
   getCameras().then((webcamDevices) => {
     store.dispatch({ type: WebCamConstants.LIST_WEBCAMS, webcamDevices });
     const state = store.getState();
-    webcamStreamSetup(webcamDevices[state.webcam.index]);
+    setWebcamStream(webcamDevices[state.webcam.index]);
   });
 }
