+++ src/action/webcam/change-webcam.js
@@ -1,10 +1,10 @@
-import { store } from '../model/store';
-import { webcamStreamSetup } from './webcam-stream-setup.js'
+import { store } from '../../model/store';
+import { setWebcamStream } from './set-webcam-stream.js'
 
 export const changeWebcam = (deviceId) => {
   const webcamsDevices = store.getState().webcamsDevices;
   const device = webcamsDevices.find((device) => {
     return (device.deviceId === deviceId)
   });
-  webcamStreamSetup(device);
+  setWebcamStream(device);
 }
