+++ src/view/kc-webcam.js
@@ -2,7 +2,7 @@ import { connect } from 'pwa-helpers'
 import { LitElement, html } from 'lit-element';
 import { store } from '../model/store.js'
 
-export class KCWebCam extends connect(store)(LitElement) {
+class KCWebCam extends connect(store)(LitElement) {
   stateChanged (state) {
     const player = this.shadowRoot.getElementById('player');
     if (player && player.srcObject !== state.webcam.stream) {
@@ -18,3 +18,5 @@ export class KCWebCam extends connect(store)(LitElement) {
 }
 
 customElements.define('kc-webcam', KCWebCam);
+
+export default KCWebCam
